+++ frontend/src/pages/dev/editor/fileHandler/csvHandler.ts
@@ -4,7 +4,7 @@ import { Polygon } from '@/mod/data/polygon/polygon';
 import type { Coord } from '@/mod/data/com/vertex';
 
 export function parseCSV(fileData: string) {
-    const nestedArrayData: string[][] = fileData.split('\n').map((line) => line.split(','));
+    const nestedArrayData: string[][] = fileData.split('\n').map((line) => line.split(';'));
     // Remove the last line due to trailing \n
     nestedArrayData.pop();
 
@@ -13,8 +13,9 @@ export function parseCSV(fileData: string) {
     for (const dataLine of nestedArrayData) {
         const id: string = dataLine[0];
         const vertices: Coord[] = [];
-        for (let idx = 1; idx < dataLine.length; idx += 2) {
-            vertices.push([parseFloat(dataLine[idx]), parseFloat(dataLine[idx + 1])]);
+        for (let idx = 1; idx < dataLine.length; idx++) {
+			const splitCoord = dataLine[idx].split(',')
+            vertices.push([parseFloat(splitCoord[0]), parseFloat(splitCoord[1])]);
         }
         const roomPolygon: Polygon = new Polygon(vertices, [70, 70, 70, 0.4]);
         const roomData: Room = new Room(0, 0, roomPolygon, id);